SILVER-TO/requirements.txt
amply==0.1.4
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.5.6
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.8.0
besos==1.3.3
bleach==3.3.0
certifi==2019.9.11
cffi==1.14.5
chardet==3.0.4
colorama==0.4.4
contexttimer==0.3.3
cplex==20.1.0.2
cycler==0.10.0
dataclasses==0.6
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.3
docloud==1.0.375
docplex==2.11.176
docutils==0.17.1
entrypoints==0.3
eppy==0.5.56
et-xmlfile==1.1.0
future==0.17.1
idna==2.8
importlib-metadata==4.5.0
inspyred==1.0.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.8.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupytext==1.11.3
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lxml==4.6.3
markdown-it-py==1.1.0
MarkupSafe==2.0.1
matplotlib==3.1.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mdit-py-plugins==0.2.8
mistune==0.8.4
more-itertools==8.8.0
multiprocess==0.70.11.1
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.4.0
numpy==1.17.0
openpyxl==2.6.3
packaging==20.9
pandas==0.24.2
pandocfilters==1.4.3
parso==0.8.2
pathos==0.2.7
pickleshare==0.7.5
Platypus-Opt==1.0.4
plotly==4.1.0
pluggy==0.13.1
ply==3.11
pox==0.2.9
ppft==1.6.6.3
prometheus-client==0.11.0
prompt-toolkit==3.0.18
PuLP==2.4
py==1.10.0
pycparser==2.20
pyDOE2==1.3.0
pydot3k==1.0.17
pyehub==1.4.2
Pygments==2.9.0
pyKriging==0.2.0
pylint==2.8.3
Pyomo==5.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.1.1
python-dateutil==2.8.1
pytz==2021.1
PyUtilib==6.0.0
pywin32==301
pywinpty==1.1.2
PyYAML==5.4.1
pyzmq==22.1.0
rbfopt==4.2.2
requests==2.22.0
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.9.0
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tinynumpy==1.2.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.4.1
